functions/tracks/lyrics_extraction/requirements.txt
beautifulsoup4
google-cloud-storage
functions-framework==3.*